security settings

+++ api/internal/controllers/post.go
@@ -2,6 +2,7 @@ package controllers
 
 import (
 	"github.com/gofiber/fiber/v2"
+	"github.com/golang-jwt/jwt/v5"
 
 	"github.com/jaum3fp/bitacora-forum/internal/dtos"
 	"github.com/jaum3fp/bitacora-forum/internal/models"
@@ -93,6 +94,29 @@ func (h *PostController) UpdatePost(c *fiber.Ctx) error {
 
 func (h *PostController) DeletePost(c *fiber.Ctx) error {
 	id := c.Params("id")
+
+	post, err := h.PostRepo.GetPost(id)
+
+	if err != nil {
+		return c.Status(fiber.StatusInternalServerError).JSON(fiber.Map{"error": err.Error()})
+	}
+
+	aTkn := c.Cookies("access-token")
+	if aTkn == "" {
+		return c.Status(fiber.StatusUnauthorized).JSON(fiber.Map{"error": "Unautorized"})
+	}
+	user := c.Locals("user").(*jwt.Token)
+	claims := user.Claims.(jwt.MapClaims)
+	username, ok := claims["username"].(string)
+
+	if !ok {
+		return c.Status(fiber.StatusUnauthorized).JSON(fiber.Map{"error": "No se ha podido identificar al usuario"})
+	}
+
+	if username != post.OwnerUsername {
+		return c.Status(fiber.StatusUnauthorized).JSON(fiber.Map{"error": "El usuario no puede realizar esta acción"})
+	}
+
 	if err := h.PostRepo.DeletePost(id); err != nil {
 		return c.Status(fiber.StatusInternalServerError).JSON(fiber.Map{"error": err.Error()})
 	}
